requirements.txt
python-telegram-bot==13.15
urllib3==1.26.20
six==1.16.0
setuptools==80.9.0
filelock==3.19.1
qrcode==8.2
pillow==11.3.0
